requirements.txt
click
pydantic
pyzmq

#CLI tools
textual

# LedDriver
# adafruit-circuitpython-neopixel


pyyaml
